refactor: Restructure the create payment modal - Added button in admin's Payment section to create new payment from source, stripe or benevity - Restructured the modal file-structure
