+++ src/components/admin/payments/create-payment/benevity/helpers/form/ExchangeRate.tsx
@@ -1,0 +1,8 @@
+import { Typography } from '@mui/material'
+
+type ExchangeRateProps = {
+  exchangeRate: number
+}
+export const ExchangeRate = ({ exchangeRate }: ExchangeRateProps) => {
+  return <Typography fontSize={17}>Курс: {exchangeRate.toFixed(2)}</Typography>
+}
