+++ src/components/admin/payments/create-payment/stripe/helpers/validations.ts
@@ -1,0 +1,4 @@
+import * as yup from 'yup'
+export const stripeInputValidation = yup.object({
+  extPaymentIntentId: yup.string().required().matches(/^ch_/, 'Невалиден номер на Страйп'),
+})
